+++ jest.config.js
@@ -1,11 +1,16 @@
 export default {
-  preset: 'ts-jest/esm',
+  preset: 'ts-jest',
   extensionsToTreatAsEsm: ['.ts'],
   testEnvironment: 'node',
   roots: ['<rootDir>/tests'],
   testMatch: ['**/*.test.ts'],
   moduleNameMapper: {
     '^(\\.{1,2}/.*)\\.js$': '$1',
+  },
+  globals: {
+    'ts-jest': {
+      diagnostics: { ignoreCodes: [151002] },
+    },
   },
   collectCoverageFrom: [
     'src/**/*.ts',